+++ mobile/android/android-components/components/lib/state/src/test/java/mozilla/components/lib/state/ext/StoreExtensionsKtTest.kt
@@ -11,6 +11,11 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LifecycleRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.ObsoleteCoroutinesApi
+import kotlinx.coroutines.channels.consumeEach
+import kotlinx.coroutines.launch
 import mozilla.components.lib.state.Store
 import mozilla.components.lib.state.TestAction
 import mozilla.components.lib.state.TestState
@@ -23,6 +28,8 @@ import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.Robolectric
+import java.util.concurrent.CountDownLatch
+import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
 class StoreExtensionsKtTest {
@@ -112,6 +119,76 @@ class StoreExtensionsKtTest {
         owner.lifecycleRegistry.markState(Lifecycle.State.DESTROYED)
         store.dispatch(TestAction.IncrementAction).joinBlocking()
         assertFalse(stateObserved)
+    }
+
+    @Test
+    @Synchronized
+    @ExperimentalCoroutinesApi // Channel
+    @ObsoleteCoroutinesApi // consumeEach
+    fun `Reading state updates from channel`() {
+        val owner = MockedLifecycleOwner(Lifecycle.State.INITIALIZED)
+
+        val store = Store(
+            TestState(counter = 23),
+            ::reducer
+        )
+
+        var receivedValue = 0
+        var latch = CountDownLatch(1)
+
+        val channel = store.broadcastChannel(owner)
+
+        GlobalScope.launch {
+            channel.openSubscription().consumeEach { state ->
+                receivedValue = state.counter
+                latch.countDown()
+            }
+        }
+
+        // Nothing received yet.
+        assertFalse(latch.await(1, TimeUnit.SECONDS))
+        assertEquals(0, receivedValue)
+
+        // Updating state: Nothing received yet.
+        store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertFalse(latch.await(1, TimeUnit.SECONDS))
+        assertEquals(0, receivedValue)
+
+        // Switching to STARTED state: Receiving initial state
+        owner.lifecycleRegistry.markState(Lifecycle.State.STARTED)
+        assertTrue(latch.await(1, TimeUnit.SECONDS))
+        assertEquals(24, receivedValue)
+        latch = CountDownLatch(1)
+
+        store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertTrue(latch.await(1, TimeUnit.SECONDS))
+        assertEquals(25, receivedValue)
+        latch = CountDownLatch(1)
+
+        store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertTrue(latch.await(1, TimeUnit.SECONDS))
+        assertEquals(26, receivedValue)
+        latch = CountDownLatch(1)
+
+        // Closing channel nothing is received anymore
+        channel.close()
+
+        store.dispatch(TestAction.IncrementAction).joinBlocking()
+        assertFalse(latch.await(1, TimeUnit.SECONDS))
+        assertEquals(26, receivedValue)
+    }
+
+    @Test(expected = IllegalArgumentException::class)
+    @ExperimentalCoroutinesApi // Channel
+    fun `Creating broadcast channel throws if lifecycle is already DESTROYED`() {
+        val owner = MockedLifecycleOwner(Lifecycle.State.DESTROYED)
+
+        val store = Store(
+            TestState(counter = 23),
+            ::reducer
+        )
+
+        store.broadcastChannel(owner)
     }
 
     @Test